requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Redis
redis>=5.0.1

# AWS S3
boto3>=1.34.34

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.3

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.1

# Logging
loguru>=0.7.2
